+++ CHANGELOG.md
@@ -2,14 +2,15 @@ CHANGELOG
 =========
 
 
-4.8, 2021-07-30
+4.8, 2021-07-31
 ---------------
-- show selection of chats to sync on online-page in case-insensitive order;
+- added shared files download for HTML export with subfolder;
 - fixed chats not being shown on importing Skype export (issue #100);
 - fixed export on Linux not using selected format (issue #101);
 - fixed certain save-dialog overwrite prompts not working on Linux;
 - fixed feedback screenshot not updating on Linux;
-- fixed help texts in advanced options dialog showing ampersands as accelerators.
+- fixed help texts in advanced options dialog showing ampersands as accelerators;
+- show selection of chats to sync on online-page in case-insensitive order.
 
 
 4.7, 2021-03-15